+++ cookbooks/chub_solr/metadata.rb
@@ -1,8 +1,8 @@
-name             'chub-solr'
+name             'chub_solr'
 maintainer       'CommerceHub Inc.'
 maintainer_email 'wgrunenwald@commercehub.com'
 license          'Apache 2.0'
-description      'Installs/Configures chub-solr'
+description      'Installs/Configures chub_solr'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version '0.2.0'